requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
httpx>=0.27.0
numpy>=1.26.0
faiss-cpu>=1.8.0
sentence-transformers>=2.6.0
pypdf>=4.2.0
pydantic>=2.6.0
flask>=3.0.0
python-dotenv>=1.0.0
